relax problem - allow one-off negative events
A one-off negative event count is very strange: it
corresponds to the assumption that an event "went
missing" somehow. There may not be a good justification
of it in a given application. But allowing negative
events is equivalent to allowing some slack in the
constraints between the observed event counts and
the contributing hidden event counts. This appears to
make the linear program much easier to solve.

This relaxation is motived by one of the relaxations
that are applied to the vehicle routing problem.